[C10D] Fix pointToPoint op Flight Recording
Fix and test issues with both coalesced and individual send/recv ops

Considered an alternate approach and then ditched it
 - alternate approach: #119757
 - reason ditched: prefer recording individual collective events inside
   coalescing region instead of just the event at the end of the region,
   which also would not have tensor sizes or opnames without additional
   state variables added

This PR instead avoids creating a work inside pointToPoint when
coalescing is active. Instead, only at endCoalescing() is a work finally
intialized and enqueued.  But it adds a record() call inside
pointToPoint() instead of creating a work, during coalescing. This
record() call picks up tensor shapes and op names.

It ALSO changes the default flight recording to record only on
enqueueWork rather than inside initWork.  This should be a safer policy
in general as we assume cuda events captured during record() will be
kept alive by work, but that lifetime is only gauranteed if the work was
enqueued.

A downside is that we have to smuggle additional tensor shape
and profilingTitle along with the Work object if we want to record it.
- Open question- how to resolve this?

The testing uncovers some odd pre-existing behavior and leaves them
alone for now. We could change some of these
- seq starts off at 1, not 0 for first op (but this is inconistent)

ghstack-source-id: 5ec6d22d9b88100f6e2b8dc6fde15c15796a4689
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120270